requirements.txt
Django>=4.2,<5
crispy-bootstrap4>=2022.1
django-crispy-forms>=2.1
django-formtools>=2.5
python-dotenv>=1.0

numpy>=1.26
pandas>=2.2
scikit-learn>=1.4
scipy>=1.11
nltk>=3.8
Unidecode>=1.3
imbalanced-learn>=0.12
imbalanced-ensemble>=0.2
xgboost>=2.0
lightgbm>=4.0
